+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.43.0] 2021-11-19
+
+- Added close button, onClose callback and ability to trigger onClose on key press to Modal component
+- Added smoother animation to Modal component
+
 ## [1.42.2] 2021-11-12
 
 - Fixed icons FarmerBook2, Global, Network, PinCheck, PinGlobal, PinPlay, PinPlus, PinUp, ProductScan, Project, Report and Team to have the same width and height